+++ app/Http/Controllers/Admin/JobRoleController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Job;
 use App\Models\JobRole;
 use App\Services\Notify;
 use App\Traits\Searchable;
@@ -83,6 +84,11 @@ class JobRoleController extends Controller
      */
     public function destroy(string $id)
     {
+        // Check if the job role is associated with any job
+        $jobExists = Job::where('job_role_id', $id)->exists();
+        if ($jobExists) {
+            return response(['message' => 'This job role is associated with some jobs, please remove them first!'], 400);
+        }
         try {
             JobRole::findOrFail($id)->delete();
             Notify::deletedNotification();
