+++ app/Http/Controllers/Admin/JobTypeController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Job;
 use App\Models\JobType;
 use App\Services\Notify;
 use App\Traits\Searchable;
@@ -83,6 +84,13 @@ class JobTypeController extends Controller
      */
     public function destroy(string $id)
     {
+
+        // check if job type is being used by any job
+        $jobExists = Job::where('job_type_id', $id)->exists();
+        if ($jobExists) {
+            return response(['message' => 'This Job Type is being used by a job, you cannot delete it!'], 400);
+        }
+
         try {
             JobType::findOrFail($id)->delete();
             Notify::deletedNotification();
